refactor: split event contracts by layer
